% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dai.R
\name{dai}
\alias{dai}
\title{dai}
\usage{
dai(n_t, t = 23, epsilon = rnorm(1, 0, 0.15), theta = c(gamma_high =
  0.439, gamma_low = 0.309, T_lag = 2.97, N_c = 276, K = 176000), DF = 600)
}
\arguments{
\item{n_t}{The current population size}

\item{t}{Number of hours between serial dilutions}

\item{epsilon}{random shock for dilution. Note: only applied
once per dilution (e.g once per function call)}

\item{theta}{parameters for the model; defaults to estimates from Dai et al; see details}

\item{DF}{Dilution factor, the environmental variable manipulated to cause the bifurcation}
}
\value{
the population size the next day, after serial dilution and growth
}
\description{
Growth of yeast population under a serial dilution regime, based on Dai et al 2012 Science paper
}
\details{
Dai et al model:

$n_{t+1} = n_t g(n_t + \\epsilon n_t, \\theta)$
$g(n_t) = n_{t+1} / n_t$

From the supplement of Dai et al (2012), Science:
"This model is based on two phases of daily
growth: a slow exponential growth phase at low cell densities, followed by a logistic growth
phase with a higher per capita growth rate at intermediate cell densities. This model has 5
parameters: T lag is the lag time before yeast cells start to grow after being transferred into new
media (the total time for daily growth is 23 hours). In the slow exponential phase, the population
grows with a constant per capita growth rate gamma_low . After the population reaches a threshold
density N_c , the subsequent logistic growth is determined by gamma_high (gamma_high > gamma_low) and the carrying
capacity K"
}
\examples{
max_days <- 30
DF <- seq(0, 2000, length=max_days) # schedule for env degredation (increased dilution)
y <- numeric(max_days)

y[1] <- 1.76e5 # initial density

for(day in 1:(max_days-1)){
y[day+1] <- dai(y[day], DF = DF[day])
}

plot(seq_along(y), y)
}

